rollback 182252c. just default kubernetes to latest stable release.
Defaulting to RCs was going a bit too far. OTOH given the increased
stability and feature consistency/completeness of upstream it doesn't
make too much sense hard-coding specific kubernetes versions.
So, and similarly to what we already do regarding CoreOS versions we
just default to whatever is the latest released kubernetes, pinning
only and when we really have too.

Signed-off-by: António Meireles <antonio.meireles@reformi.st>