Avoid download binaries on subsequent vagrant up
Instead of downloading the binaries everytime vagrant up is done, do
a ConditionalPath to check for binaries. This also prevents the kubectl
from getting downloaded if there is already a kubectl in /usr/local/bin

Please note that this does not check the version of kubectl but just the
existence.